+++ src/components/Assistant/bubble.tsx
