+++ src/components/Assistant/index.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useRef, useState } from "react";
 
 import { GPTMessage, GPTMessageSchema } from "@/lib/chatbot/validators";
-import MessengerEmbed from "../Header/messengerembed/messengerembed";
 
 import Bubble from "./bubble";
 import Message from "./message";
+import MessengerEmbed from "../MessengerChat";
 
 import styles from "./assistant.module.css";
 
@@ -70,7 +70,7 @@ function Assistant() {
     setChatOpen(!chatOpen);
   };
 
-  //Switch to MessengerEmbed
+  // Switch to MessengerEmbed
   const handleButtonClick = () => {
     setShowChatbot(false);
   };
